todo-list-api/pyproject.toml
[project]
name = "todo-list-api"
version = "0.1.0"
description = "A RESTful API to allow users to manage their to-do list."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-limiter>=4.1.1",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
